consolidate adaptive throttler impls
